Rename wwlib to WWLib
