+++ Code/Commando/DlgMPConnect.cpp
@@ -40,7 +40,7 @@
 #include "campaign.h"
 #include "cnetwork.h"
 #include "resource.h"
-#include <wwdebug\wwdebug.h>
+#include <WWDebug\wwdebug.h>
 #include "dlgmainmenu.h"
 #include "gamespyadmin.h"
 #include "specialbuilds.h"
