+++ Code/WWDebug/CMakeLists.txt
@@ -7,7 +7,7 @@ set(SOURCES
 add_library(WWDebug STATIC ${SOURCES})
 target_include_directories(WWDebug PUBLIC
     $<BUILD_INTERFACE:
-        ${PROJECT_SOURCE_DIR}/Code/wwdebug
+        ${PROJECT_SOURCE_DIR}/Code/WWDebug
     >
     PRIVATE
         ${PROJECT_SOURCE_DIR}/Code/wwlib
@@ -20,7 +20,7 @@ if(REBEL_ENABLE_EDITOR)
     add_library(WWDebugE STATIC ${SOURCES})
     target_include_directories(WWDebugE PUBLIC
         $<BUILD_INTERFACE:
-            ${PROJECT_SOURCE_DIR}/Code/wwdebug
+            ${PROJECT_SOURCE_DIR}/Code/WWDebug
         >
         PRIVATE
             ${PROJECT_SOURCE_DIR}/Code/wwlib
